python_app/requirements.txt
psycopg2-binary
sqlalchemy
boto3
tqdm
pillow
numpy
pandas